pyproject.toml
[tool.poetry]
name = "supabase-full-text-search"
version = "0.1.0"
description = ""
authors = ["Andrew Smith <a.smith@silentworks.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
supabase = "^2.4.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
